update mesh.py in drop_element method
The method drop_element don't drop the attribute "name" and attribute "space" of elements in the name and space list of mesh class. This cause some problems with "mesh.dump2inp" method, because then "len(self.labels)" ist not equal to "len(self.name)" and so the for loop "for i in xrange(len(self.labels)): elements[self.name[i]].append(self.labels[i])" stops before all elements are listed in the out string. The drop of attribute "space" has no influence to the 'dump2inp' method, but it also should be changed .